% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentModel.R
\name{backStepIter}
\alias{backStepIter}
\title{backwards removes \beta_i in model
  Y       \sim N(X \beta + X \gamma, \sigma_Y^2 I_{n \times n})
  \gamma  \sim N(u, \tau^2 I_{p \times p})
  u       = [u_1,\ldots,u_{n_c}]
  u_i     \sim N(\mu_0, \sigma^2_u 1 %*% 1 ^T)}
\usage{
backStepIter(
  Y,
  X,
  model = c("simple", "constant"),
  betas,
  critval = qnorm(1 - (1 - 0.95)/2),
  nc = NULL,
  return_full = FALSE,
  SVDX = NULL
)
}
\arguments{
\item{Y}{- (n x 1) observations}

\item{X}{- (n x p) covariates}

\item{model}{- (string)
simple   - simpliefed model (sigma^2_u = 0)
constant - the full model}

\item{critval}{- (double) value to minimum accepted level}

\item{nc}{- (int) number of chromosones}

\item{return_full}{- (bool) return either betas or betas and params}
}
\value{
betas      - (p x 1) vector of the beta values
}
\description{
backwards removes \beta_i in model
  Y       \sim N(X \beta + X \gamma, \sigma_Y^2 I_{n \times n})
  \gamma  \sim N(u, \tau^2 I_{p \times p})
  u       = [u_1,\ldots,u_{n_c}]
  u_i     \sim N(\mu_0, \sigma^2_u 1 %*% 1 ^T)
}
