% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModel.R
\name{backward_GeneMixObj}
\alias{backward_GeneMixObj}
\title{backward selection of find}
\usage{
backward_GeneMixObj(GeneMixObj, find, crit = mBIC2_GeneMixObj, silent = T)
}
\description{
backward selection of find
}
