% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentModel.R
\name{BICmixed.R}
\alias{BICmixed.R}
\title{computing the BIC for the model in loglikSetup function
Not implimented for \sigma_U
 @param theta (3 x 1) log(\sigma_Y), log(\tau), log(\sigma_U)
 @param likobj (list) contaning:
                Yu     (n x 1) the data (modifed by U in the SVD(X))
                Xu     (n x k + 1) the covariates, for \beta (modifed by U in the SVD(X))
                H      (n x n_c) covariance matrix
                HHt    (n x n)
                SVDX   (list) singular value decomposition of X
 @param tau_fixed - (bool) assuming that \tau,\sigma are fixed
 @param Xu        - (k x 1) covariates supplied (optional)
 @param find      - (m x 1) number of ocv
 @param C         - (double) if non NULL it is a MBIC}
\usage{
BICmixed.R(theta, likObj, tau_fixed = F, Xu = NULL, find = NULL, C = NULL)
}
\description{
computing the BIC for the model in loglikSetup function
Not implimented for \sigma_U
 @param theta (3 x 1) log(\sigma_Y), log(\tau), log(\sigma_U)
 @param likobj (list) contaning:
                Yu     (n x 1) the data (modifed by U in the SVD(X))
                Xu     (n x k + 1) the covariates, for \beta (modifed by U in the SVD(X))
                H      (n x n_c) covariance matrix
                HHt    (n x n)
                SVDX   (list) singular value decomposition of X
 @param tau_fixed - (bool) assuming that \tau,\sigma are fixed
 @param Xu        - (k x 1) covariates supplied (optional)
 @param find      - (m x 1) number of ocv
 @param C         - (double) if non NULL it is a MBIC
}
