% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CholUpdate}
\alias{CholUpdate}
\title{Update a Cholesky factor R of Q=R'R
to R of Q+XX'= R'R}
\usage{
CholUpdate(R, U)
}
\arguments{
\item{R}{upper triangel matrix}

\item{x}{is a matrix}
}
\description{
Update a Cholesky factor R of Q=R'R
to R of Q+XX'= R'R
}
