% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModel.R
\name{estimateGeneMix}
\alias{estimateGeneMix}
\title{estimate GeneMixObj}
\usage{
estimateGeneMix(GeneMixObj, theta0 = NULL, find = NULL, restricted = T)
}
\arguments{
\item{GeneMixObj}{- gene mixture object}

\item{theta0}{- known intial parameters deafult null}
}
\description{
estimate GeneMixObj
}
