% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Figure1-data.R
\docType{data}
\name{Figure1}
\alias{Figure1}
\title{Figure one in Ghost QTL ...}
\format{
An object of class \code{list} of length 3.
}
\usage{
See Figures/Figure1.R
}
\description{
Data for creating Figure 1 in
Data for Figure in Ghost QTL and hotspots in experimental crosses - novel approach for modeling polygenic effects
origin: don't remeber
}
\details{
Figure 1 in the article
}
\keyword{datasets}
