% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModel.R
\name{GeneBackward}
\alias{GeneBackward}
\title{A more general backward model for geneObject}
\usage{
GeneBackward(GeneMixObj, dupl = NULL, restricted = T, C = 4, silent = F)
}
\arguments{
\item{C}{- mBIO2 parameter}
}
\description{
A more general backward model for geneObject
}
