% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModel.R
\name{GeneForward}
\alias{GeneForward}
\title{A more general forward model for geneObject}
\usage{
GeneForward(
  GeneMixObj,
  candidates = NULL,
  dupl = NULL,
  C = 4,
  restricted = T,
  silent = F
)
}
\description{
A more general forward model for geneObject
}
