% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentModel.R
\name{getBeta}
\alias{getBeta}
\title{gets the ML estimate of the covariates given coeffients
 @param theta (3 x 1) log(\sigma_Y), log(\tau), log(\sigma_U)
 @param likobj (list) contaning:
                Yu     (n x 1) the data (modifed by U in the SVD(X))
                Xu     (n x k + 1) the covariates, for \beta (modifed by U in the SVD(X))
                H      (n x n_c) covariance matrix
                HHt    (n x n)
                SVDX   (list) singular value decomposition of X
 @param restricted (bool) use restricted maximum likelihood
 @param Xu   - (k x 1) covariates supplied (optional)}
\usage{
getBeta(theta, likObj, Xu = NULL)
}
\description{
gets the ML estimate of the covariates given coeffients
 @param theta (3 x 1) log(\sigma_Y), log(\tau), log(\sigma_U)
 @param likobj (list) contaning:
                Yu     (n x 1) the data (modifed by U in the SVD(X))
                Xu     (n x k + 1) the covariates, for \beta (modifed by U in the SVD(X))
                H      (n x n_c) covariance matrix
                HHt    (n x n)
                SVDX   (list) singular value decomposition of X
 @param restricted (bool) use restricted maximum likelihood
 @param Xu   - (k x 1) covariates supplied (optional)
}
