% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentModel.R
\name{getCov_old}
\alias{getCov_old}
\title{Get the \mu,v covariances (not joint)}
\usage{
getCov_old(theta, likObj)
}
\description{
@param theta (3 x 1) log(\sigma_Y), log(\tau), log(\sigma_U)
 @param likobj (list) contaning:
                Yu     (n x 1) the data (modifed by U in the SVD(X))
                Xu     (n x k + 1) the covariates, for \beta, \mu_0 (modifed by U in the SVD(X))
                H      (n x n_c) covariance matrix
                HHt    (n x n)
                SVDX   (list) singular value decomposition of X
}
