% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentModel.R
\name{getSigmaUY}
\alias{getSigmaUY}
\title{Returns the covariance matrix of t(U)%*%Y
 if model is simple return the diagonal vector
 @param theta (3 x 1) \sigma_Y, \tau, \sigma_U
 @param likobj (list) contaning:
                Yu     (n x 1) the data (modifed by U in the SVD(X))
                Xu     (n x k + 1) the covariates, for \beta, \mu_0 (modifed by U in the SVD(X))
                H      (n x n_c) covariance matrix
                HHt    (n x n)
                SVDX   (list) singular value decomposition of X}
\usage{
getSigmaUY(theta, likObj)
}
\description{
Returns the covariance matrix of t(U)%*%Y
 if model is simple return the diagonal vector
 @param theta (3 x 1) \sigma_Y, \tau, \sigma_U
 @param likobj (list) contaning:
                Yu     (n x 1) the data (modifed by U in the SVD(X))
                Xu     (n x k + 1) the covariates, for \beta, \mu_0 (modifed by U in the SVD(X))
                H      (n x n_c) covariance matrix
                HHt    (n x n)
                SVDX   (list) singular value decomposition of X
}
