% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentModel.R
\name{gettvaluesCond2}
\alias{gettvaluesCond2}
\title{Caculates the t-values for X_i in :
  Y       \sim N(X_i \beta_i + X \gamma, \sigma_Y^2 I_{n \times n})
  \gamma  \sim N(u, \tau^2 I_{p \times p})
  u       = [u_1,\ldots,u_{n_c}]
  u_i     \sim N(\mu_0, \sigma^2_u 1 %*% 1 ^T)}
\usage{
gettvaluesCond2(
  index,
  Y,
  X,
  model = c("simple", "constant"),
  betas = NULL,
  nc = NULL,
  theta = NULL,
  X_rem = NULL,
  SVDX = NULL
)
}
\arguments{
\item{index}{- (p_1 x 1) vector of position where to calc t-values}

\item{Y}{- (n x 1) observations}

\item{X}{- (n x p) covariates}

\item{model}{- (string)
simple   - simpliefed model (sigma^2_u = 0)
constant - the full model}

\item{nc}{- (int) number of chromosones}

\item{theta}{- (k x 1) parameters of the model if NULL estimated from data}

\item{XSVD}{- (list) singular value decomposition of X}
}
\value{
t - (p x 1) vector of the t-values
}
\description{
Caculates the t-values for X_i in :
  Y       \sim N(X_i \beta_i + X \gamma, \sigma_Y^2 I_{n \times n})
  \gamma  \sim N(u, \tau^2 I_{p \times p})
  u       = [u_1,\ldots,u_{n_c}]
  u_i     \sim N(\mu_0, \sigma^2_u 1 %*% 1 ^T)
}
