% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentModelR.R
\name{iterativeForwardSelectR}
\alias{iterativeForwardSelectR}
\title{Iterative procedure for selecting covariates uder base model:
  Y       \sim N(X_k\beta + X \gamma, \sigma_Y^2 I_{n \times n})
  \gamma  \sim N(u, \tau^2 I_{p \times p})
  u       = [u_1,\ldots,u_{n_c}]
  u_i     \sim N(\mu_0, \sigma^2_u 1 %*% 1 ^T)
In e}
\usage{
iterativeForwardSelectR(
  Y,
  X,
  model = c("simple", "constant"),
  kmax = NULL,
  nc = NULL,
  X_rem = NULL,
  critval = qnorm(1 - (1 - 0.95)/2),
  likelihood_method = FALSE,
  silent = 1
)
}
\arguments{
\item{Y}{- (n x 1) observations}

\item{X}{- (n x p) covariates}

\item{model}{- (string)
simple   - simpliefed model (sigma^2_u = 0)
constant - the full model}

\item{nc}{- (int) number of chromosones}

\item{X_rem}{- (p< x 1) number of columns X that should not be checked (like duplicates)}

\item{critval}{- (double) value to minimum accepted level}

\item{likelihood_method}{- (bool) use highest likelihood to select coeff}

\item{silent}{(int) 0 - no output
1 - result of each loop
2 - with stars for iteration}

\item{Kmax}{- (int) maximum number of iterations}
}
\value{
beta - (p x 1) vector with the selected betas (and values)
}
\description{
Iterative procedure for selecting covariates uder base model:
  Y       \sim N(X_k\beta + X \gamma, \sigma_Y^2 I_{n \times n})
  \gamma  \sim N(u, \tau^2 I_{p \times p})
  u       = [u_1,\ldots,u_{n_c}]
  u_i     \sim N(\mu_0, \sigma^2_u 1 %*% 1 ^T)
In e
}
