% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentModel.R
\name{loglikSetup}
\alias{loglikSetup}
\title{setup for computing the loglikelihood for the model
Assumes equally spaced covariates
  Y       \sim N(X_k\beta + X \gamma, \sigma_Y^2 I_{n \times n})
  \gamma  \sim N(u, \tau^2 I_{p \times p})
  u       = [u_1,\ldots,u_{n_c}]
  u_i     \sim N(\mu_0, \sigma^2_u 1 %*% 1 ^T)
 @param Y     (n x 1) the data
 @param X     (n x p) the gene matrixs
 @param Xk    (n x k) the covariates (for \beta) if no covariates set NULL
 @param nc    (int - matrix)   number of chromosones, if null simpler model}
\usage{
loglikSetup(Y, X, Xk = NULL, nc = NULL, SVDX = NULL)
}
\description{
@param SVDX  (list) output of SVD(X)
}
