% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModel.R
\name{mBIC2_GeneMixObj}
\alias{mBIC2_GeneMixObj}
\title{mBIC
takes a an object and a find list
and caculates the mBIC value}
\usage{
mBIC2_GeneMixObj(GeneMixObj, find, const = 4)
}
\arguments{
\item{find}{- (k x 1) known non zero genes}
}
\description{
mBIC
takes a an object and a find list
and caculates the mBIC value
}
