% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModel.R
\name{mixedModelForwardBackward}
\alias{mixedModelForwardBackward}
\title{fitting a mixed model with selection on X usign forward selection}
\usage{
mixedModelForwardBackward(
  GeneMixObj,
  markers,
  dCM = 1,
  dupl = NULL,
  liberal = TRUE,
  qval = NULL,
  alpha = 0.05,
  estParam = T
)
}
\arguments{
\item{markers}{- (m x 1) length of the m chromosnes}

\item{dCM}{- (double) distance in centiMorgan}

\item{dupl}{- (c x 1) duplicated rows of X (found in GeneMixObj$LikObj$X)}
}
\description{
fitting a mixed model with selection on X usign forward selection
}
