% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModel.R
\name{mixedModelGetBeta}
\alias{mixedModelGetBeta}
\title{fitting a mixed model with selection on X usign forward selection}
\usage{
mixedModelGetBeta(GeneMixObj)
}
\description{
fitting a mixed model with selection on X usign forward selection
}
