% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneModel.R
\name{SetupGeneMix}
\alias{SetupGeneMix}
\title{Base class for gene model
  Y       \sim N(X_k\beta + X \gamma, \sigma_Y^2 I_{n \times n})
  \gamma  \sim N(u, \tau^2 I_{p \times p})
  u       = [u_1,\ldots,u_{n_c}]
  u_i     \sim N(\mu_0, \sigma^2_u 1 %*% 1 ^T)}
\usage{
SetupGeneMix(
  formula,
  data,
  X,
  tauOff = FALSE,
  sigmaMuOff = TRUE,
  meanMuOff = FALSE,
  nc = NULL,
  SVDX = NULL,
  testTau = FALSE,
  nPC = 0
)
}
\arguments{
\item{formula}{- (formula)   the covariates formula}

\item{data}{- (data.frame) for covariates (X_k) these should not include X}

\item{X}{- (n x p) the gene covariates}

\item{tauOff}{- (bool) fix \tau = 0,}

\item{sigmaMuOff}{- (bool) \sigma_u = 0, if sigmaMuOff=FALSE needs to include nc}

\item{meanMuOff}{- (bool) \mu_0,= 0,}

\item{nc}{- (int) number of chromosones, if null simpler model}

\item{SVDX}{- (obj) singular value decomposition object of X}

\item{testTau}{- (bool) should we try to turn off tau?}

\item{nPC}{- (int) how many PC to use in fixed effect}
}
\description{
Base class for gene model
  Y       \sim N(X_k\beta + X \gamma, \sigma_Y^2 I_{n \times n})
  \gamma  \sim N(u, \tau^2 I_{p \times p})
  u       = [u_1,\ldots,u_{n_c}]
  u_i     \sim N(\mu_0, \sigma^2_u 1 %*% 1 ^T)
}
