% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationGene.R
\name{simulation_GeneData}
\alias{simulation_GeneData}
\title{simulation types}
\usage{
simulation_GeneData(
  type,
  k = 10,
  n = 500,
  L = 100,
  tau = 0.02,
  mu0 = 0.005,
  sigma = 1,
  betas = NULL,
  betasVal = NULL
)
}
\arguments{
\item{type}{(int) 1,2,3
1 - pure mixed effect
2 - varying effect over chromosones
3 - varying effect over each chromosone}

\item{k}{(int) number of chromosone}

\item{n}{(int) number of observations}

\item{L}{(int) length of chromosne (i.e number locus)}

\item{tau}{(double) beta variation}

\item{sigma}{(double) standard deviation of measurement error}

\item{betas}{(k*L) which betas are nonzero}

\item{betasVal}{(k*L or 1x1) which are the values of betas}

\item{mu}{(double) baseline beta}
}
\description{
simulation types
}
