% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationGene.R
\name{simulation_GeneDataSelection}
\alias{simulation_GeneDataSelection}
\title{simulation of Xs through selection
input}
\usage{
simulation_GeneDataSelection(beta, X, ngen, p, lambda_cross)
}
\arguments{
\item{beta}{- (n x 1) coeffients}

\item{X}{- (N x n x 2) original coeffients}

\item{ngen}{- (int)  number of generations to simulate}

\item{p}{- (N x 1) probability of generating offspring}

\item{lambda_cross}{- (double) intensty of number of crossing}
}
\description{
simulation of Xs through selection
input
}
