% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationGene.R
\name{simulation_mixing}
\alias{simulation_mixing}
\title{mixing two populations}
\usage{
simulation_mixing(X1, X2, lambda_cross, generations = 1)
}
\arguments{
\item{X1}{- (N1 x n x 2) original X pop1}

\item{X2}{- (N2 x n x 2) original X pop2}

\item{lambda_cross}{- (double) intensty of number of crossing}

\item{ngen}{- (int)  number of generations to simulate}
}
\description{
mixing two populations
}
