% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{summarizeFind.new}
\alias{summarizeFind.new}
\title{calculate power, FP, FDR, FWER, distance and beta error}
\usage{
summarizeFind.new(find, beta, beta.true, qtl.pos, maxCM = 15)
}
\arguments{
\item{find}{- (q x 1) list of position discovered}

\item{beta}{- (q x 1) list of the coeffient of the discoveries}

\item{beta.true}{- (m x 1) vector of the true coeffients}

\item{qtl.pos}{- (m x 1) vecotr of the true positions}

\item{maxCM}{- (int) minimum abs dist coutning as a find}
}
\description{
calculate power, FP, FDR, FWER, distance and beta error
}
