% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentModel.R
\name{whiteData}
\alias{whiteData}
\title{whiting selected vectors}
\usage{
whiteData(
  GeneMixObj,
  theta = NULL,
  XOut = T,
  YOut = T,
  XmOut = T,
  Xextra = NULL
)
}
\arguments{
\item{GeneMixObj}{-}

\item{theta}{- (vec) parameters for variance model}

\item{XmOut}{- (bool) whitning covariates}

\item{Xextra}{- (n x p_0) extra covariates to whiten (should be in U coordinates (i.e X- SVD basis))}

\item{Xout}{- (bool) whitning entire gene matrix}

\item{Yout}{- (bool) whitning output}
}
\description{
whiting selected vectors
}
