% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WoodburySpecial}
\alias{WoodburySpecial}
\title{WoodburySpecial
Computes the inverse of D + UU^T where
D is a diagonal matrix
U is a low dimensional matrix}
\usage{
WoodburySpecial(D, U)
}
\arguments{
\item{D}{is a vector of the diagonal of D}

\item{U}{is the low rank matrix}
}
\description{
WoodburySpecial
Computes the inverse of D + UU^T where
D is a diagonal matrix
U is a low dimensional matrix
}
